move from winapi to windows-sys

+++ src/lookup.rs
@@ -5,8 +5,13 @@ use std::str;
 #[cfg(unix)]
 use libc::{NI_NUMERICSERV, SOCK_STREAM};
 
+/*
 #[cfg(windows)]
 use winapi::shared::ws2def::{NI_NUMERICSERV, SOCK_STREAM};
+*/
+
+#[cfg(windows)]
+use windows_sys::Win32::Networking::WinSock::{NI_NUMERICSERV, SOCK_STREAM};
 
 use addrinfo::{getaddrinfo, AddrInfoHints};
 use nameinfo::getnameinfo;
@@ -16,7 +21,7 @@ use nameinfo::getnameinfo;
 /// Returns an iterator of IP Addresses, or an `io::Error` on failure.
 pub fn lookup_host(host: &str) -> io::Result<Vec<IpAddr>> {
     let hints = AddrInfoHints {
-        socktype: SOCK_STREAM,
+        socktype: SOCK_STREAM as i32,
         ..AddrInfoHints::default()
     };
 
@@ -37,7 +42,7 @@ pub fn lookup_host(host: &str) -> io::Result<Vec<IpAddr>> {
 /// Returns the hostname as a String, or an `io::Error` on failure.
 pub fn lookup_addr(addr: &IpAddr) -> io::Result<String> {
     let sock = (*addr, 0).into();
-    match getnameinfo(&sock, NI_NUMERICSERV) {
+    match getnameinfo(&sock, NI_NUMERICSERV as i32) {
         Ok((name, _)) => Ok(name),
         Err(e) => {
             reload_dns_nameserver();
